Added ability to create skins in subdirectories of existing suites
When using the "New Skin..." command, if a path is entered instead of a name,
the skin that is created will be placed in this path (relative to the skins
folder) and named after the last folder name. The @Resources folder will be
created in the correct location if it's not already there. See the README for
instructions.